requirements.txt
Django==1.4.3
South==0.7.6
argparse==1.2.1
boto==2.8.0
distribute==0.6.24
dj-database-url==0.2.1
django-grappelli==2.4.3
django-storages==1.1.6
django-tastypie==0.9.12
gunicorn==0.17.2
mimeparse==0.1.3
psycopg2==2.4.6
python-dateutil==2.1
six==1.2.0
wsgiref==0.1.2
